reorganize into package format
